Split trace models by responsibility
